refactor(global): global 패키지 분리
