+++ src/main/java/nextstep/subway/global/domain/BaseEntity.java
@@ -1,4 +1,4 @@
-package nextstep.subway.global;
+package nextstep.subway.global.domain;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
