+++ src/main/java/nextstep/subway/global/domain/FindFailedException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.global;
+package nextstep.subway.global.domain;
 
 public class FindFailedException extends RuntimeException {
