+++ src/main/java/nextstep/subway/member/application/MemberService.java
@@ -1,6 +1,6 @@
 package nextstep.subway.member.application;
 
-import nextstep.subway.global.FindFailedException;
+import nextstep.subway.global.domain.FindFailedException;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import nextstep.subway.member.dto.MemberRequest;
