+++ src/main/java/nextstep/subway/member/domain/Member.java
@@ -1,7 +1,7 @@
 package nextstep.subway.member.domain;
 
 import nextstep.subway.auth.application.AuthorizationException;
-import nextstep.subway.global.BaseEntity;
+import nextstep.subway.global.domain.BaseEntity;
 import org.apache.commons.lang3.StringUtils;
 
 import javax.persistence.Entity;
