+++ src/main/java/nextstep/subway/path/domain/PathFindFailedException.java
@@ -1,6 +1,6 @@
 package nextstep.subway.path.domain;
 
-import nextstep.subway.global.BusinessException;
+import nextstep.subway.global.domain.BusinessException;
 
 public class PathFindFailedException extends BusinessException {
